upgrade moby client, rename versioninfo to version
