+++ Makefile
@@ -4,9 +4,9 @@ REPO_PATH := github.com/projecteru2/core
 REVISION := $(shell git rev-parse HEAD || unknown)
 BUILTAT := $(shell date +%Y-%m-%dT%H:%M:%S)
 VERSION := $(shell git describe --tags $(shell git rev-list --tags --max-count=1))
-GO_LDFLAGS ?= -s -X $(REPO_PATH)/versioninfo.REVISION=$(REVISION) \
-			  -X $(REPO_PATH)/versioninfo.BUILTAT=$(BUILTAT) \
-			  -X $(REPO_PATH)/versioninfo.VERSION=$(VERSION)
+GO_LDFLAGS ?= -s -X $(REPO_PATH)/version.REVISION=$(REVISION) \
+			  -X $(REPO_PATH)/version.BUILTAT=$(BUILTAT) \
+			  -X $(REPO_PATH)/version.VERSION=$(VERSION)
 
 grpc:
 	cd ./rpc/gen/; protoc --go_out=plugins=grpc:. core.proto
