+++ rpc/rpc.go
@@ -13,7 +13,7 @@ import (
 	"github.com/projecteru2/core/cluster"
 	pb "github.com/projecteru2/core/rpc/gen"
 	"github.com/projecteru2/core/types"
-	"github.com/projecteru2/core/versioninfo"
+	"github.com/projecteru2/core/version"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
 )
@@ -31,9 +31,9 @@ type Vibranium struct {
 // Info show core info
 func (v *Vibranium) Info(ctx context.Context, opts *pb.Empty) (*pb.CoreInfo, error) {
 	return &pb.CoreInfo{
-		Version:       versioninfo.VERSION,
-		Revison:       versioninfo.REVISION,
-		BuildAt:       versioninfo.BUILTAT,
+		Version:       version.VERSION,
+		Revison:       version.REVISION,
+		BuildAt:       version.BUILTAT,
 		GolangVersion: runtime.Version(),
 		OsArch:        fmt.Sprintf("%s/%s", runtime.GOOS, runtime.GOARCH),
 	}, nil
